# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "librero-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/libros": {
      "get": {
        "tags": [
          "crate"
        ],
        "operationId": "get_libros",
        "responses": {
          "200": {
            "description": "Todos los libros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Libro"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate"
        ],
        "operationId": "add_libro",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddLibro"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Libro successfully added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Libro"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddLibro": {
        "type": "object",
        "required": [
          "isbn"
        ],
        "properties": {
          "isbn": {
            "type": "string"
          }
        }
      },
      "Libro": {
        "type": "object",
        "required": [
          "isbn",
          "title",
          "author"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "cover_path": {
            "type": "string",
            "nullable": true
          },
          "isbn": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      }
    }
  }
}